+++ src/methods/$enumerate/$enumerate.js
@@ -1,8 +1,13 @@
-import range from '../range/range';
-import $zip from '../$zip/$zip';
+import { $iterableCurry } from '../../internal/$iterable';
+import { range } from '../range/range';
+import { $zip } from '../$zip/$zip';
+import { $wrap } from '../$wrap/$wrap';
 
 export function $enumerate(iterable, start = 0) {
-  return $zip(range(start, Infinity, 1), iterable);
+  return $zip([$wrap(range(start, Infinity)), iterable]);
 }
 
-export default $enumerate;
+export default $iterableCurry($enumerate, {
+  minArgs: 0,
+  maxArgs: 1,
+});