+++ src/methods/$enumerate/enumerate.mjs
@@ -6,9 +6,14 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import range from '../range/range';
-import zip from '../$zip/zip';
+import { iterableCurry } from '../../internal/iterable';
+import { range } from '../range/range';
+import { zip } from '../$zip/zip';
+import { wrap } from '../$wrap/wrap';
 export function enumerate(iterable, start = 0) {
-  return zip(range(start, Infinity, 1), iterable);
+  return zip([wrap(range(start, Infinity)), iterable]);
 }
-export default enumerate;
+export default iterableCurry(enumerate, {
+  minArgs: 0,
+  maxArgs: 1,
+});